aider: Changed home route in server.py to serve the client/index.html file instead of returning "Hello, World!".
# Aider chat conversation:

ASSISTANT: To check if the HTML file is being served correctly, we need to look at the server-side code. In this case, it's the `server/server.py` file.

Currently, the server is set up to return "Hello, World!" when the root URL is accessed. To serve the `client/index.html` file instead, we need to modify the home route in `server/server.py`.

Here is the change:

```python
server/server.py
<<<<<<< HEAD
@app.route('/', methods=['GET'])
def home():
    return "Hello, World!"
=======
from flask import send_from_directory

@app.route('/', methods=['GET'])
def home():
    return send_from_directory('client', 'index.html')
>>>>>>> updated
```

This change will make Flask serve the `client/index.html` file when the root URL is accessed.

Please add the `server/server.py` file to the chat so I can make the changes.