aider: Added requests and beautifulsoup4 to requirements.txt.
# Aider chat conversation:

ASSISTANT: Great! To implement the `crawler.py` file, we need to create a function that will crawl from the start page to the finish page on Wikipedia.

We will use the `requests` and `bs4` (BeautifulSoup) libraries to make HTTP requests and parse the HTML of the Wikipedia pages.

We will also use a breadth-first search (BFS) algorithm to find the shortest path between the start page and the finish page. BFS is an algorithm for traversing or searching tree or graph data structures. It starts at the tree root (or some arbitrary node of a graph, sometimes referred to as a 'search key') and explores the neighbor nodes at the present depth prior to moving on to nodes at the next depth level.

Here are the steps we will follow:

1. Create a queue and enqueue the start page.
2. Create a set to keep track of visited pages.
3. While the queue is not empty, dequeue a page.
4. If this page is the finish page, we have found a path.
5. Otherwise, for each link on the page that leads to another Wikipedia page, if we have not visited that page before, enqueue it and mark it as visited.

We will also need to add `requests` and `bs4` to the `requirements.txt` file.

Here are the changes we need to make:

1. Modify `server/requirements.txt` to add `requests` and `bs4`.
2. Modify `server/crawler.py` to implement the `find_path` function.

Let's start with the `requirements.txt` file. Here is the *edit block* for that:

```python
server/requirements.txt
<<<<<<< HEAD
Flask==1.1.2
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
itsdangerous==1.1.0
=======
Flask==1.1.2
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
itsdangerous==1.1.0
requests==2.25.1
beautifulsoup4==4.9.3
>>>>>>> updated
```

After you approve this change, I will provide the *edit block* for the `crawler.py` file.