aider: Refactored `find_path` function to implement breadth-first search algorithm in the crawler.
# Aider chat conversation:

USER: ok, I want to make a crawler that uses breadth first search

ASSISTANT: Alright, to implement a breadth-first search (BFS) in the crawler, we will need to modify the `find_path` function in the `server/crawler.py` file.

The BFS algorithm works by visiting the start node first, then visiting all the nodes adjacent to the start node, then visiting all nodes adjacent to those nodes, and so on. In the context of a web crawler, the nodes are web pages and the edges are links between pages.

Here are the steps we will follow:

1. Create a queue and enqueue the start page.
2. Create a set to keep track of visited pages.
3. While the queue is not empty, dequeue a page.
4. If this page is the finish page, we have found a path.
5. Otherwise, for each page linked from this page that has not been visited, enqueue the page and mark it as visited.

We will also need a function to get all the links from a page. This function will make a request to the page, parse the HTML to find all the links, and return a list of links.

Here is the code change:

```python
server/crawler.py
<<<<<<< HEAD
def find_path(start_page, finish_page):
    return [start_page, 'Some intermediate page', finish_page]
=======
import requests
from bs4 import BeautifulSoup

def get_links(page_url):
    response = requests.get(page_url)
    soup = BeautifulSoup(response.text, 'html.parser')
    links = [a['href'] for a in soup.find_all('a', href=True)]
    return links

def find_path(start_page, finish_page):
    queue = [(start_page, [start_page])]
    visited = set()

    while queue:
        (vertex, path) = queue.pop(0)
        for next in set(get_links(vertex)) - visited:
            if next == finish_page:
                return path + [next]
            else:
                visited.add(next)
                queue.append((next, path + [next]))
    return []
>>>>>>> updated
```

This code uses the BeautifulSoup library to parse HTML. If this library is not already installed, you will need to add it to your `requirements.txt` file.