aider: Refactored get_links() function in server/crawler.py to prepend 'https://www.wikipedia.org' to each href if it starts with '/wiki/'.
# Aider chat conversation:

USER:  * Running on http://127.0.0.1:5000/ (Press CTRL+C to quit)
127.0.0.1 - - [14/Oct/2023 20:47:33] "GET / HTTP/1.1" 200 -
[2023-10-14 20:47:39,673] ERROR in server: Error occurred: Invalid URL '/wiki/Enigma_rotor_details': No schema supplied. Perhaps you meant http:///wiki/Enigma_rotor_details?
127.0.0.1 - - [14/Oct/2023 20:47:39] "POST /find_path HTTP/1.1" 500 -

ASSISTANT: The error message indicates that the URL provided to the `requests.get()` function in the `get_links()` function of `server/crawler.py` is invalid because it lacks a schema (http or https). This is likely because the href attribute of the anchor tags in the HTML being parsed are relative URLs (e.g., '/wiki/Enigma_rotor_details') rather than absolute URLs.

To fix this, we need to modify the `get_links()` function in `server/crawler.py` to prepend 'https://www.wikipedia.org' to each href if it starts with '/wiki/'.

Here is the change:

```python
server/crawler.py
<<<<<<< HEAD
    response = requests.get(page_url)
    soup = BeautifulSoup(response.text, 'html.parser')
    links = [a['href'] for a in soup.find_all('a', href=True)]
=======
    response = requests.get(page_url)
    soup = BeautifulSoup(response.text, 'html.parser')
    links = ['https://www.wikipedia.org' + a['href'] if a['href'].startswith('/wiki/') else a['href'] for a in soup.find_all('a', href=True)]
>>>>>>> updated
```

This change will ensure that all URLs passed to the `requests.get()` function are absolute URLs, which should resolve the error.