aider: Add print statement to output the links that the crawler is following.
# Aider chat conversation:

USER: for logging purposes, can we output to console the links that the crawler is following?

ASSISTANT: Yes, we can add a print statement in the `find_path()` function in `server/crawler.py` to output the links that the crawler is following. Here is the change:

```python
server/crawler.py
<<<<<<< HEAD
    while queue:
        (vertex, path) = queue.pop(0)
        for next in set(get_links(vertex)) - visited:
            if next == finish_page:
                return path + [next]
            else:
                visited.add(next)
                queue.append((next, path + [next]))
=======
    while queue:
        (vertex, path) = queue.pop(0)
        for next in set(get_links(vertex)) - visited:
            print(f"Following link: {next}")
            if next == finish_page:
                return path + [next]
            else:
                visited.add(next)
                queue.append((next, path + [next]))
>>>>>>> updated
```

This change will print a message to the console each time the crawler follows a link.