aider: Updated find_path function to print out messages when a link is added to the queue and when the finish page is found.
# Aider chat conversation:

ASSISTANT: The order in which links are processed in the `find_path` function in `server/crawler.py` is determined by the order in which they are returned by the `get_links` function. The `get_links` function returns links in the order they are found in the HTML of the page, which may not correspond to the order they are displayed on the page due to the way the HTML is structured.

If you want to process links in a different order, you would need to modify the `get_links` function to sort the links in the desired order before returning them. However, keep in mind that the order in which links are processed can significantly affect the performance of the `find_path` function. Processing links in the order they are found in the HTML is generally the most efficient approach, as it minimizes the amount of time spent waiting for network requests to complete.

If you want to see the order in which links are processed, you could modify the `find_path` function to print out each link as it is processed. Here is an *edit block* for that change:

```python
server/crawler.py
<<<<<<< HEAD
    while queue:
        (vertex, path) = queue.pop(0)
        for next in set(get_links(vertex)) - visited:
            print(f"Following link: {next}")
            if next == finish_page:
                return path + [next]
            else:
                visited.add(next)
                queue.append((next, path + [next]))
=======
    while queue:
        (vertex, path) = queue.pop(0)
        for next in set(get_links(vertex)) - visited:
            print(f"Following link: {next}")
            if next == finish_page:
                print(f"Found finish page: {next}")
                return path + [next]
            else:
                print(f"Adding link to queue: {next}")
                visited.add(next)
                queue.append((next, path + [next]))
>>>>>>> updated
```

This change will print out a message each time a link is added to the queue, as well as when the finish page is found.