aider: Changed the JavaScript code to display the path line by line on the webpage.
# Aider chat conversation:

USER: I wanted the path output line by line but that is not what is happening

ASSISTANT: I understand that you want the path to be displayed line by line on the webpage. This can be achieved by modifying the JavaScript code in the `client/main.js` file.

Currently, the path is joined with `<br>` which is an HTML tag for line break. However, the `innerHTML` property of the `path` element is not interpreting it as HTML but as a string. We can fix this by creating a new paragraph element for each path and appending it to the `path` element.

Here are the changes:

```javascript
client/main.js
<<<<<<< HEAD
    .then(data => {
        document.getElementById('path').innerHTML = data.path.join('<br>');
    });
=======
    .then(data => {
        var pathElement = document.getElementById('path');
        pathElement.innerHTML = ''; // clear previous path
        data.path.forEach(function(page) {
            var para = document.createElement("p");
            var node = document.createTextNode(page);
            para.appendChild(node);
            pathElement.appendChild(para);
        });
    });
>>>>>>> updated
```

This change will create a new paragraph for each page in the path and append it to the `path` element, effectively displaying the path line by line.