aider: Refactored the code to build an HTML string and set it as the innerHTML of the path element.
# Aider chat conversation:

USER: there is sth fundamentally wrong with the way we try to turn the path to html. whatever you propose, nothing changes the output, it is always a simple line.

ASSISTANT: I see. It seems like the issue might be with the way the HTML is being updated in the JavaScript code. Let's try a different approach. Instead of creating and appending HTML elements using JavaScript, we can build an HTML string and set it as the innerHTML of the path element. This approach is less efficient for large updates, but it should work fine for our use case.

Here are the changes:

```javascript
client/main.js
<<<<<<< HEAD
    .then(data => {
        var pathElement = document.getElementById('path');
        pathElement.innerHTML = ''; // clear previous path
        var ul = document.createElement('ul');
        data.path.forEach(function(page) {
            var li = document.createElement("li");
            var a = document.createElement("a");
            a.href = page;
            a.textContent = decodeURIComponent(page);
            li.appendChild(a);
            ul.appendChild(li);
        });
        pathElement.appendChild(ul);
    });
=======
    .then(data => {
        var pathElement = document.getElementById('path');
        pathElement.innerHTML = ''; // clear previous path
        var html = '<ul>';
        data.path.forEach(function(page) {
            html += '<li><a href="' + page + '">' + decodeURIComponent(page) + '</a></li>';
        });
        html += '</ul>';
        pathElement.innerHTML = html;
    });
>>>>>>> updated
```

This change will build an HTML string with a list item and a link for each page in the path. The link's URL is the page URL, and the link's text is the decoded page name. The HTML string is then set as the innerHTML of the path element.