aider: Added query string to the end of the `main.js` file URL to force the browser to download the new version.
# Aider chat conversation:

ASSISTANT: It sounds like the issue might be related to caching. Browsers often cache static files like JavaScript and CSS to improve performance. This means that even if you change the file on the server, the browser might still use the old, cached version.

One way to force the browser to download the new version of the file is to add a query string to the end of the file URL. The browser treats URLs with different query strings as different resources, so it will download the new version of the file instead of using the cached version.

Here's how you can do it:

```html
client/index.html
<<<<<<< HEAD
    <script src="/static/main.js"></script>
=======
    <script src="/static/main.js?v=2"></script>
>>>>>>> updated
```

This change adds a query string `?v=2` to the end of the `main.js` file URL. You can increment the number after `v=` each time you make a change to the `main.js` file to ensure the browser downloads the new version.

Please note that this is a workaround for development and might not be suitable for a production environment. In a production environment, you would typically use a build tool or a server configuration to handle cache busting.